Separate the base api resource into it's own helper file, and write tests
