+++ backend/my_app/blueprints/api.py
@@ -1,63 +1,11 @@
-from flask import request, Blueprint
-from flask_restful import Resource, abort, Api
-from inflection import underscore
+from flask import Blueprint
+from flask_restful import Api
 
-from backend.my_app import db
 from backend.my_app.models import model_classes
+from backend.my_app.helpers.base_resource import BaseResource
 
 api_blueprint = Blueprint('api', __name__)
 api = Api(api_blueprint)
-
-
-class BaseResource(Resource):
-    class_model = None
-    class_schema = None
-    url = None
-
-    @classmethod
-    def get_url(cls):
-        if cls.url:
-            return cls.url
-        return '/api/{}'.format(underscore(cls.__name__).replace('_resource',
-                                                                 ''))
-
-    def get(self):
-        schema_instance = self.class_schema()
-        get_obj = self.class_model.query.get(request.args['id'])
-
-        if get_obj:
-            schema_result = schema_instance.dump(get_obj)
-            return schema_result.data
-        else:
-            return abort(404, message="No {} object found with id {}".format(
-                self.class_model.__name__, request.args['id']
-            ))
-
-    def post(self):
-        schema_instance = self.class_schema()
-        result = schema_instance.load(request.json, db.session)
-        post_obj = result.data
-        db.session.add(post_obj)
-        db.session.commit()
-        schema_result = schema_instance.dump(post_obj)
-        return schema_result.data
-
-    def put(self):
-        schema_instance = self.class_schema()
-        result = schema_instance.load(request.json, db.session)
-        put_obj = result.data
-        db.session.commit()
-        schema_result = schema_instance.dump(put_obj)
-        return schema_result.data
-
-    def delete(self):
-        schema_instance = self.class_schema()
-        result = schema_instance.load(request.json, db.session)
-        delete_obj = result.data
-        db.session.delete(delete_obj)
-        db.session.commit()
-        schema_result = schema_instance.dump(delete_obj)
-        return schema_result.data
 
 
 for model in model_classes:
@@ -67,6 +15,7 @@ for model in model_classes:
         (BaseResource,),
         {
             "class_schema": model.__marshmallow__,
+            "out_schema": model.__out_marshmallow__,
             "class_model": model
         }
     )
